refactor: ReservationTime의 JdbcTemplate에 의존하던 부분 제거
